+++ jwst/extract_1d/tests/test_interpolate_response.py
@@ -1,6 +1,8 @@
 """
 Test for extract_1d.interpolate_response
 """
+import math
+
 import numpy as np
 
 from jwst.datamodels import SlitModel
@@ -17,10 +19,16 @@ def test_interpolate_response():
     tbl1 = np.array(list(zip(wl1, response)), dtype=slit.relsens.dtype)
     tbl2 = np.array(list(zip(wl2, response)), dtype=slit.relsens.dtype)
 
-    r1 = extract.interpolate_response(wavelength, tbl1, verbose=False)
-    assert len(r1) == len(wavelength)
-    assert r1[1] == 13.0
+    reciprocal_r1 = extract.interpolate_response(wavelength, tbl1,
+                                                 verbose=False)
+    assert len(reciprocal_r1) == len(wavelength)
+    assert math.isclose(1. / reciprocal_r1[1], 13.0, rel_tol=1.e-6)
+    assert reciprocal_r1[0] == 0.       # because wavelength is out of range
+    assert reciprocal_r1[4] == 0.
 
-    r2 = extract.interpolate_response(wavelength, tbl2, verbose=False)
-    assert len(r2) == len(wavelength)
-    assert r2[1] == 35.0
+    reciprocal_r2 = extract.interpolate_response(wavelength, tbl2,
+                                                 verbose=False)
+    assert len(reciprocal_r2) == len(wavelength)
+    assert math.isclose(1. / reciprocal_r2[1], 35.0, rel_tol=1.e-6)
+    assert reciprocal_r2[0] == 0.       # because wavelength is out of range
+    assert reciprocal_r2[4] == 0.